refactor: Ticket과 Lotto 모델 분리
