+++ src/js/model/Lotto.js
@@ -1,25 +1,4 @@
-function Ticket() {
-  const init = () => {
-    this.numbers = getLottoNumber();
-    this.profit = 0;
-  };
-
-  const getRandomNumber = (min, max) => {
-    return Math.floor(Math.random() * (max - min + 1)) + min;
-  };
-
-  const getLottoNumber = () => {
-    const lottoNumbers = new Set();
-
-    while (lottoNumbers.size < 6) {
-      lottoNumbers.add(getRandomNumber(1, 45));
-    }
-
-    return [...lottoNumbers].sort((a, b) => a - b);
-  };
-
-  init();
-}
+import Ticket from './Ticket.js';
 
 export default function Lotto(lottoNums) {
   const init = () => {
@@ -31,7 +10,7 @@ export default function Lotto(lottoNums) {
   };
 
   this.putLottoNumbers = () => {
-    //
+    return this.tickets.map((ticket) => ticket.numbers);
   };
 
   init();
